+++ base/application/api_hooks/api_hooks.py
@@ -6,21 +6,21 @@ pack_user(AuthUser) -> [dict, None]:
         - return users data as dictionary
 check_password_is_valid(password) -> bool:
         - check for password validation
-register_user(id_user, username, password, data) -> [dict, str, convertible to string, None]:
+register_user(id_user, username, password, data, request_handler=None) -> [dict, str, convertible to string, None]:
         - register user on system
         - populate auth_users and users tables here
-pre_register_user(username, password, data) -> [None]:
+pre_register_user(username, password, data, request_handler=None) -> [None]:
         - process user's data before user registration
-post_register_process(id_user, username, password, data, session_token) -> [dict, None]:
+post_register_process(id_user, username, password, data, session_token, request_handler=None) -> [dict, None]:
         - process user's data after user registration
 user_exists(username, password, data, handler) -> [User object]
         - check if username exists in the system
 check_username_and_password(username, password, Auth_user) -> [bool]
         - check username / password match
-pre_login_process(Auth_user, json_data) -> [dict, str, None]
+pre_login_process(Auth_user, json_data, request_handler=None) -> [dict, str, None]
         - pre login data processing
         - on error raise PreLoginError
-post_login_process(Auth_user, json_data, token) -> [dict, str, None]
+post_login_process(Auth_user, json_data, token, request_handler=None) -> [dict, str, None]
         - after login processing
         - on error raise PostLoginError
 save_hash(hash_data) -> [dict, str]
@@ -33,7 +33,7 @@ pre_logout_process(Auth_user) -> [dict, None]
         - pre logout data processing
 post_logout_process(Auth_user, session_token) -> [dict, None]
         - post logout data processing
-check_user(Auth_user) -> [dict]
+check_user(Auth_user, request_handler=None) -> [dict]
         - check user process
 pre_check_user(Auth_user) -> [dict]
         - process user before check user process
@@ -104,7 +104,7 @@ def check_password_is_valid(password):
     return True
 
 
-def register_user(id_user, username, password, data):
+def register_user(id_user, username, password, data, request_handler=None):
     """
     Save user into database
     :param id_user: database user id
@@ -182,7 +182,7 @@ def check_username_and_password(username, password, user):
     return True
 
 
-def check_user(auth_user):
+def check_user(auth_user, request_handler=None):
     """
     Check logged user and return it's data.
     On error raise CheckUserError exception